fix: filter out the --user-data-dir arg from flags
This flag is generated by Puppeteer, and points to a temporary directory
if the `userDataDir` option is not supplied to `launch()`. However, the
temporary directory is created on the host machine and may not work
correctly in the container. It also changes how Puppeteer closes the
browser. See #1 for more information.